docs(metrics): clarify single metrics metadata & publish
